included Music container
