+++ src/App.js
@@ -3,7 +3,8 @@ import { Route, Switch, Redirect } from 'react-router-dom';
 
 import Detail from './components/Detail/Detail'
 import PageNotFound from './components/PageNotFound/404'
-import Music from './components/Music/Music';
+import MusicContainer from './containers/MusicContainer';
+
 import PlayList from './components/PlayList/PlayList';
 
 import './App.css';
@@ -13,7 +14,7 @@ class App extends Component {
     return (
       <div className="App">
         <Switch>
-          <Route path="/" exact component={Music} />
+          <Route path="/" exact component={MusicContainer} />
           <Route path="/detail/:artist/:album" component={Detail} />
           <Route path="/playlist" component={PlayList} />
           <Route path="/404" component={PageNotFound} />
