+++ src/components/Music/Music.js
@@ -1,11 +1,7 @@
 import React, { Component, Fragment } from 'react';
+
 import Search from '../Search/Search';
 import SongsList from '../SongsList/SongsList';
-import styled from 'styled-components';
-import { connect } from 'react-redux';
-
-import { search, closeModal, changeQS } from '../../actions';
-import { IModalMessage } from '../Modal/modalMessages';
 import { Button } from '../Styled/Button';
 import { Header } from '../Styled/Header';
 import Modal from '../Modal/Modal';
@@ -79,32 +75,4 @@ class Music extends Component {
     }
 }
 
-
-const mapDispatchToProps = dispatch => (
-    {
-        onSearch: (queryString, pageNumber) => {
-            dispatch(search(queryString, pageNumber))
-        },
-        onCloseModal: () => {
-            dispatch(closeModal())
-        },
-        onChangeQueryString: (queryString) => {
-            dispatch(changeQS(queryString))
-        },
-    }
-)
-
-const mapStateToProps = state => (
-    {
-        searching: state.searching,
-        results: state.searchResult,
-        showModal: state.showModal,
-        queryString: state.queryString,
-        anySelected: state.selection.length > 0,
-        modalMessage: state.modalMessage
-    }
-)
-
-export default connect(mapStateToProps, mapDispatchToProps)(
-    withFooter(Music, 30)
-);
+export default withFooter(Music, 30);
